removed extra keymaps for encoder, and updated encoder code
