+++ keyboards/fancytech/fancyalice66/fancyalice66.c
@@ -60,7 +60,7 @@ void rgb_matrix_indicators_user(void)
 
 #endif
 
-#ifdef ENCODER_ENABLE
+#if defined ENCODER_ENABLE && !defined VIA_ENABLE
 bool encoder_update_kb(uint8_t index, bool clockwise) {
     if (!encoder_update_user(index, clockwise)) { 
         return false;
