+++ keyboards/fancytech/fancyalice66/keymaps/default/keymap.c
@@ -18,7 +18,7 @@
 const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
 
     /*
-     *    ┌───┬───┬───┬───┬───┬───┬───┐   ┌───┬───┬───┬───┬───┬───┬───────┐  ┌───┐ X  X
+     *    ┌───┬───┬───┬───┬───┬───┬───┐   ┌───┬───┬───┬───┬───┬───┬───────┐  ┌───┐ Vol-  Vol+
      *    │Esc│ 1 │ 2 │ 3 │ 4 │ 5 │ 6 │   │ 7 │ 8 │ 9 │ 0 │ - │ = │Backsp │  │Del│
      *   ┌┴───┴┬──┴┬──┴┬──┴┬──┴┬──┴┬──┘ ┌─┴─┬─┴─┬─┴─┬─┴─┬─┴─┬─┴─┬─┴─┬─────┴┐ └┬──┴┐
      *   │ Tab │ Q │ W │ E │ R │ T │    │ Y │ U │ I │ O │ P │ [ │ ] │  \   │  │PgD│
