--- keyboards/fancytech/fancyalice66/keymaps/default/rules.mk
@@ -1,1 +1,0 @@
-ENCODER_ENABLE = no
