+++ keyboards/fancytech/fancyalice66/keymaps/via/rules.mk
@@ -1,3 +1,2 @@
-ENCODER_ENABLE = no
 VIA_ENABLE = yes
 LTO_ENABLE = yes
