--- keyboards/fancytech/fancyalice66/keymaps/via_encoder/keymap.c
@@ -1,79 +1,0 @@
-/* Copyright 2022 chent7 <https://github.com/chent7>
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 2 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-#include QMK_KEYBOARD_H
-
-const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
-
-    /*
-     *    ┌───┬───┬───┬───┬───┬───┬───┐   ┌───┬───┬───┬───┬───┬───┬───────┐  ┌───┐ Vol-  Vol+
-     *    │Esc│ 1 │ 2 │ 3 │ 4 │ 5 │ 6 │   │ 7 │ 8 │ 9 │ 0 │ - │ = │Backsp │  │MUT│
-     *   ┌┴───┴┬──┴┬──┴┬──┴┬──┴┬──┴┬──┘ ┌─┴─┬─┴─┬─┴─┬─┴─┬─┴─┬─┴─┬─┴─┬─────┴┐ └┬──┴┐
-     *   │ Tab │ Q │ W │ E │ R │ T │    │ Y │ U │ I │ O │ P │ [ │ ] │  \   │  │PgD│
-     *  ┌┴─────┼───┼───┼───┼───┼───┤    └┬──┴┬──┴┬──┴┬──┴┬──┴┬──┴┬──┴──────┴┐ └┬──┴┐
-     *  │ Caps │ A │ S │ D │ F │ G │     │ H │ J │ K │ L │ ; │ ' │  Enter   │  │PgU│
-     * ┌┴──────┴┬──┴┬──┴┬──┴┬──┴┬──┴┐    └─┬─┴─┬─┴─┬─┴─┬─┴─┬─┴─┬─┴───────┬─┬┴──┼───┘
-     * │ Shift  │ Z │ X │ C │ V │ B │      │ N │ M │ , │ . │ / │  Shift  │ │Up │
-     * ├─────┬──┴───┼───┴─┬─┴───┴─┬─┴─┐  ┌─┴───┴───┴┬──┴──┬┴───┴───────┬─┴─┼───┼───┐
-     * │ Ctl │      │ Alt │       │FN │  │          │ Alt │            │Lft│Dwn│Rgt│
-     * └─────┘      └─────┴───────┴───┘  └──────────┴─────┘            └───┴───┴───┘
-     */
-
-    [0] = LAYOUT(
-        KC_ESC,  KC_1,    KC_2,    KC_3,    KC_4,    KC_5,    KC_6,    KC_7,    KC_8,    KC_9,    KC_0,    KC_MINS, KC_EQL,  KC_BSPC,          KC_MUTE, KC_VOLU, KC_VOLD,
-        KC_TAB,  KC_Q,    KC_W,    KC_E,    KC_R,    KC_T,             KC_Y,    KC_U,    KC_I,    KC_O,    KC_P,    KC_LBRC, KC_RBRC, KC_BSLS, KC_PGUP,
-        KC_CAPS, KC_A,    KC_S,    KC_D,    KC_F,    KC_G,             KC_H,    KC_J,    KC_K,    KC_L,    KC_SCLN, KC_QUOT, KC_ENT,           KC_PGDN,
-        KC_LSFT, KC_Z,    KC_X,    KC_C,    KC_V,    KC_B,             KC_N,    KC_M,    KC_COMM, KC_DOT,  KC_SLSH, KC_RSFT,          KC_UP,
-        KC_LCTL,          KC_LALT, KC_SPC,                    MO(1),   KC_SPC,                    KC_RALT,                   KC_LEFT, KC_DOWN, KC_RGHT
-    ),
-
-    [1] = LAYOUT(
-        KC_TILD, KC_F1,   KC_F2,   KC_F3,   KC_F4,   KC_F5,   KC_F6,   KC_F7,   KC_F8,   KC_F9,   KC_F10,  KC_F11,  KC_F12,  RESET,            KC_PSCR, _______, _______,
-        _______, _______, _______, _______, _______, _______,          _______, _______, _______, _______, _______, RGB_SAI, RGB_SAD, RGB_TOG, RGB_HUI,
-        _______, _______, _______, _______, _______, _______,          _______, _______, _______, _______, _______, _______, _______,          RGB_HUD,
-        _______, _______, _______, _______, _______, _______,          _______, _______, _______, _______, _______, _______,          RGB_VAI,
-	    _______,          _______, _______,                   _______, _______,                   RGB_MOD,                   _______, RGB_VAD, _______
-    ),
-
-    [2] = LAYOUT(
-        _______, _______, _______, _______, _______, _______, _______,          _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
-        _______, _______, _______, _______, _______, _______,          _______, _______, _______, _______, _______, _______, _______, _______, _______,
-        _______, _______, _______, _______, _______, _______,          _______, _______, _______, _______, _______, _______, _______,          _______,
-        _______, _______, _______, _______, _______, _______,          _______, _______, _______, _______, _______, _______,          _______,
-	    _______,          _______, _______,                   _______, _______,                   _______,                   _______, _______, _______
-    ),
-
-    [3] = LAYOUT(
-        _______, _______, _______, _______, _______, _______, _______,          _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
-        _______, _______, _______, _______, _______, _______,          _______, _______, _______, _______, _______, _______, _______, _______, _______,
-        _______, _______, _______, _______, _______, _______,          _______, _______, _______, _______, _______, _______, _______,          _______,
-        _______, _______, _______, _______, _______, _______,          _______, _______, _______, _______, _______, _______,          _______,
-	    _______,          _______, _______,                   _______, _______,                   _______,                   _______, _______, _______
-    )
-};
-
-#ifdef ENCODER_ENABLE
-bool encoder_update_user(uint8_t index, bool clockwise) {
-  if (index == 0) {
-    if (clockwise) {
-        tap_code(dynamic_keymap_get_keycode(biton32(layer_state), 4, 12));
-    } else {
-        tap_code(dynamic_keymap_get_keycode(biton32(layer_state), 4, 11));
-    }
-  }
-  return true;
-}
-#endif
