--- keyboards/fancytech/fancyalice66/keymaps/via_encoder/rules.mk
@@ -1,2 +1,0 @@
-VIA_ENABLE = yes
-LTO_ENABLE = yes
